chore: green the CI alongside the entitySearch fix
PR #11 surfaced that CI has been red on main since 4/16 with pre-existing
issues unrelated to the entitySearch bug. Per reviewer request, bundling
the cleanup into this PR so the fix doesn't land on a red branch.

## Changes

### Lint (ruff) — 45 errors → 0

`ruff check . --fix` applied. All mechanical:
- Unused imports removed (typing.Callable / Iterable / Optional / Any /
  Dict / List; unittest.mock.MagicMock; pytest; pathlib.Path;
  NotificationTransformResult; LegacyDynatraceV1Client)
- Import blocks sorted / formatted (I001) across ~12 test files
- 1 extraneous `f` prefix removed (agents/java.py:21)

No semantic changes.

### Test dependency

`hypothesis>=6.100.0` added to requirements.txt. `tests/unit/
test_invariants.py` uses `from hypothesis import HealthCheck, given,
settings; from hypothesis import strategies as st` but the package was
missing from requirements, so `test` jobs on Python 3.9–3.12 all failed
with `ModuleNotFoundError: No module named 'hypothesis'`.

### Typecheck (mypy) — 2 errors → 0

- `migration/audit.py:222` — explicit `bool(...)` cast on the equality
  comparison (returned `Any` from an `Any == Any` expression)
- `migration/report.py:70` — explicit `float(...)` cast on the mean
  division (returned `Any / int` — pyright/mypy can't narrow)

No behavioral changes; both sites were already returning the correct
runtime type.

## Local verification

- `ruff check .` → All checks passed!
- `mypy compiler/ migration/ validators/ config/` → no issues in 21 source files
- `pytest tests/` → 1,271 passed, 14 skipped, 0 failed

+++ migration/report.py
@@ -67,7 +67,7 @@ class ConversionReport:
         ]
         if not scored:
             return None
-        return sum(scored) / len(scored)
+        return float(sum(scored) / len(scored))
 
     def summary(self) -> Dict:
         """Return aggregate statistics about conversion results."""
